Move formatValue function into UI and add the relevant tests
